Rename Execution trace class
