--- barretenberg/cpp/src/barretenberg/execution_trace/CMakeLists.txt
@@ -1,1 +1,0 @@
-barretenberg_module(execution_trace plonk_honk_shared)
