+++ barretenberg/cpp/src/barretenberg/trace_to_polynomials/trace_to_polynomials.hpp
